Mostly finished Spring Security after tutorial

+++ src/main/java/com/roomy/models/UserEntity.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private int Id;
 
     @Column(unique = true)
     private String userName;
